rename pages folder to views
