+++ src/app/app-routing.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { LoginPageComponent } from './core/auth/pages/login-page/login-page.component';
-import { LogoutComponent } from './core/auth/pages/logout/logout.component';
-import { RegisterPageComponent } from './core/auth/pages/register-page/register-page.component';
+import { LoginPageComponent } from './core/auth/views/login-page/login-page.component';
+import { LogoutComponent } from './core/auth/views/logout/logout.component';
+import { RegisterPageComponent } from './core/auth/views/register-page/register-page.component';
 import { AuthGuard } from './core/guards/auth.guard';
 import { GuestGuard } from './core/guards/guest.guard';
-import { DashboardPageComponent } from './modules/dashboard/pages/dashboard-page/dashboard-page.component';
+import { DashboardPageComponent } from './modules/dashboard/views/dashboard-page/dashboard-page.component';
 import { AuthLayoutComponent } from './core/layouts/auth-layout/auth-layout.component';
 import { MainLayoutComponent } from './core/layouts/main-layout/main-layout.component';
 import { PageNotFoundComponent } from './shared/pages/page-not-found/page-not-found.component';
