+++ src/app/core/auth/auth.module.ts
@@ -4,9 +4,9 @@ import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
 
-import { LoginPageComponent } from './pages/login-page/login-page.component';
-import { LogoutComponent } from './pages/logout/logout.component';
-import { RegisterPageComponent } from './pages/register-page/register-page.component';
+import { LoginPageComponent } from './views/login-page/login-page.component';
+import { LogoutComponent } from './views/logout/logout.component';
+import { RegisterPageComponent } from './views/register-page/register-page.component';
 
 @NgModule({
   declarations: [LoginPageComponent, LogoutComponent, RegisterPageComponent],
