+++ src/app/core/auth/views/logout/logout.component.ts
