+++ src/app/core/auth/views/register-page/register-page.component.ts
