+++ src/app/modules/dashboard/dashboard.module.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 
-import { DashboardPageComponent } from './pages/dashboard-page/dashboard-page.component';
+import { DashboardPageComponent } from './views/dashboard-page/dashboard-page.component';
 
 @NgModule({
   declarations: [DashboardPageComponent],
